[Refactor] 3주차 - Controller 부분 return 값 Json 으로 수정
