+++ contents/todoListAPI/yeongjae/todolist/todolist/src/main/java/com/example/todolist/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package com.example.todolist;
+package com.example.todolist.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
