+++ contents/todoListAPI/yeongjae/todolist/todolist/src/main/java/com/example/todolist/controller/MemberController.java
@@ -1,6 +1,5 @@
 package com.example.todolist.controller;
 
-import com.example.todolist.repository.MemberRepository;
 import com.example.todolist.service.MemberService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
@@ -8,9 +7,6 @@ import org.springframework.stereotype.Controller;
 @Controller
 @RequiredArgsConstructor
 public class MemberController {
-
-    private final MemberService memberService;
-
 
 
 }
