--- contents/todoListAPI/yeongjae/todolist/todolist/src/main/java/com/example/todolist/domain/Member.java
@@ -1,26 +1,0 @@
-package com.example.todolist.domain;
-
-import lombok.Getter;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.OneToMany;
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@Getter
-public class Member {
-
-    @Id
-    @GeneratedValue
-    private Long id;
-
-    private String name;
-    private String email;
-
-    @OneToMany(mappedBy = "member")
-    private List<Todo> todoList = new ArrayList<>();
-
-}
