+++ contents/todoListAPI/yeongjae/todolist/todolist/src/main/java/com/example/todolist/repository/MemberRepository.java
@@ -1,8 +1,8 @@
 package com.example.todolist.repository;
 
-import com.example.todolist.domain.Member;
+import com.example.todolist.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MemberRepository extends JpaRepository<Member, Long> {
+public interface MemberRepository extends JpaRepository<User, Long> {
 
 }
