+++ contents/todoListAPI/yeongjae/todolist/todolist/src/main/java/com/example/todolist/repository/TodoRepository.java
@@ -1,7 +1,7 @@
 package com.example.todolist.repository;
 
-import com.example.todolist.domain.Todo;
+import com.example.todolist.domain.Task;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TodoRepository extends JpaRepository<Todo, Long> {
+public interface TodoRepository extends JpaRepository<Task, Long> {
 }
