+++ contents/todoListAPI/yeongjae/todolist/todolist/src/main/java/com/example/todolist/service/MemberService.java
@@ -1,6 +1,6 @@
 package com.example.todolist.service;
 
-import com.example.todolist.domain.Member;
+import com.example.todolist.domain.User;
 import com.example.todolist.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -12,8 +12,8 @@ public class MemberService {
     private final MemberRepository memberRepository;
 
     @Transactional
-    public Long join(Member member) {
-        memberRepository.save(member);
-        return member.getId();
+    public Long join(User user) {
+        memberRepository.save(user);
+        return user.getId();
     }
 }
