+++ contents/todoListAPI/yeongjae/todolist/todolist/src/main/java/com/example/todolist/service/TodoService.java
@@ -1,6 +1,7 @@
 package com.example.todolist.service;
 
-import com.example.todolist.domain.Todo;
+import com.example.todolist.domain.Task;
+import com.example.todolist.dto.TaskDto;
 import com.example.todolist.repository.TodoRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -15,8 +16,8 @@ public class TodoService {
     private final TodoRepository todoRepository;
 
     @Transactional
-    public void writeTodo(Todo todo) {
-        todoRepository.save(todo);
+    public void writeTodo(TaskDto taskDto) {
+        todoRepository.save(taskDto.changeEntity(taskDto));
     }
 
     @Transactional
@@ -24,7 +25,7 @@ public class TodoService {
         todoRepository.deleteById(id);
     }
 
-    public List<Todo> findAll() {
+    public List<Task> findAll() {
         return todoRepository.findAll();
     }
 }
